lcs-preinstall: add parameters related to active directory integration
Adding these parameters is a pre-requisite for the rest of the work
on delegating authentication to an active directory server. The remaining
patches will be sent for review on the list before being committed. This
one however is harmless and only sets the foundation required by storing
the corresponding parameters.

git-svn-id: https://svn.tice.ac-caen.fr/svn/LcsSE3/LCS/LCS2.0ndg/sources/trunk@3637 1012e355-7652-49d3-8fd6-1f9b0b464cea