Switch to debhelper compat level 9 and use “dh” helper
Detailed change:
* debian/compat: change 4 (or 5) to 9
* debian/control: build-depends to debhelper >= 9
* debian/rules: use simplified form with dh and override targets

Rationale:
The debhelper suite of tools evolve to implement the latest specifications
of the Debian policy. To get the best result, it's recommended to use the
latest compatibility level. That level is currently '9' and is supported
by versions of debhelper greater than 9 (which is available in Wheezy
too).

The "dh" tool is provided by debhelper since version 7 and ensures that
you calls all the dh_* helper tools in the correct order. It lets you
replace some of those calls by a corresponding "override_dh_*" target
in debian/rules (when you want to pass extra options, or when you want
to execute custom actions before of after the dh_ helper script).

When you bump the compatibility level some of the dh_ tools may change
behaviour and it's recommended to read "man 7 debhelper" to be aware of
those differences.