Update debian/copyright to use copyright format 1.0
The debian/copyright file ought to document the licence of the various
files provided in the source package. It is supposed to document the
upstream license but in the case of the LCS, there is no distinction
between upstream authors and package maintainers so you should
keep those files up-to-date with copyright notice and appropriate
license statement.

Debian standardized the format of this file to be associate clearly
a license to each provided file and to be able to make automatic analysis
of licence usage. It is thus recommended to create debian/copyright
files following this format:
https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/

This commit updates the copyright files to use the new format and puts in
light some missing license statements, and possible usage of non-free
files, particularly in lcs-security.

Good development practice would be to include a small header in each
source file referring to a top-level LICENSE file that would include the
precise license. See
https://www.softwarefreedom.org/resources/2012/ManagingCopyrightInformation.html
for a good explanation.