Switch source format to "3.0 (native)"
Debian introduced newer source package formats to overcome some
limitations of the initial 1.0 format. It's highly recommended to use the
new formats by default.

For native packages (where the package maintainer is also the upstream
author), the format is called "3.0 (native)" and has the following
benefits:
- it supports xz compression of the tarball
- it automatically ignores VCS-specific files like .svn or .gitignore
  and other common temporary files (.swp of vim, etc.).

This commit adds debian/source/format to request usage of the new format.