Drop sbin/lcs-security-config.sh which is an unused empty file.
And its mere presence was resulting in lintian warnings:
W: lcs-security: executable-not-elf-or-script usr/sbin/lcs-security-config.sh

It was marked as executable but did not contain anything to execute.

And also:
W: lcs-security: script-with-language-extension usr/sbin/lcs-security-config.sh

Because it's not a good idea to use an extension that indicates in which
language the script was written in. It makes it harder to rewrite it later
with another language.