Add ${misc:Depends} and drop ${shlibs:Depends} to/from dependencies
The debhelper tools sometimes have to add a dependency to achieve policy
compliance. In order to achieve this they inject their dependency in
a substitution variable called ”misc:Depends”. For this to be effective,
this substitution variable must be present in the Depends line of
debian/control. This commit adds the variable where appropriate.

On the opposite, the substitution variable “shlibs:Depends” is generated
by dpkg-shlibdeps and is only useful when the package provides ELF
binaries which architecture all packages never have. This substitution
variable is thus not needed and produces noise during build because
dpkg-gencontrol will complain about an unknown variable. This commits
drops any reference to this inappropriate substitution variable.

This fixes the following lintian warnings:
W: lcs-apache2ssl source: debhelper-but-no-misc-depends lcs-apache2ssl
W: lcs-preinstall source: debhelper-but-no-misc-depends lcs-preinstall

Here's the long description of this warning:
N:
N:   The source package uses debhelper, but it does not include
N:   ${misc:Depends} in the given binary package's debian/control entry.
N:   Any debhelper command may add dependencies to ${misc:Depends} that are
N:   required for the work that it does, so recommended best practice is to
N:   always add ${misc:Depends} to the dependencies of each binary package
N:   if debhelper is in use.
N:
N:   Refer to the debhelper(7) manual page for details.
N:
N:   Severity: normal, Certainty: possible
N:
N:   Check: debhelper, Type: source
N: