Add lintian overrides to ignore dir-or-file-in-var-www
Lintian complains about the fact that files are put into /var/www.
This policy rules makes sense for a general purpose server but
in the case of the LCS, we the server setup is well under
control and it's perfectly safe to use this directory as long as
we configure our virtual hosts to point to this directory.

It would be a massive work to change to another directory and
with no significant benefits, so we purposedly ignore this
lintian error by adding a “lintian override”.

The long description of the lintian error is the following.
E: lcs-security: dir-or-file-in-var-www var/www/Admin/
N:
N:   Debian packages should not install files under /var/www. This is not
N:   one of the /var directories in the File Hierarchy Standard and is
N:   under the control of the local administrator. Packages should not
N:   assume that it is the document root for a web server; it is very
N:   common for users to change the default document root and packages
N:   should not assume that users will keep any particular setting.
N:
N:   Packages that want to make files available via an installed web server
N:   should instead put instructions for the local administrator in a
N:   README.Debian file and ideally include configuration fragments for
N:   common web servers such as Apache.
N:
N:   As an exception, packages are permitted to create the /var/www
N:   directory due to its past history as the default document root, but
N:   should at most copy over a default file in postinst for a new install.
N:   In this case, please add a Lintian override.
N:
N:   Refer to Filesystem Hierarchy Standard (The /var Hierarchy) for
N:   details.
N:
N:   Severity: serious, Certainty: possible
N:
N:   Check: files, Type: binary, udeb
N: