Setup /home/rssh and /root/.bashrc in postinst and not via packaged files
Packages are not allowed to ship files in /home and /root. Those should
be under the control of the respective users (and administrator). When
specific setup needs to be done, it's best done as one-time operation
in the postinst script. That way the users have the possibility to
override the changes made by the package. We do the required setup
only once during initial installation so that a package upgrade will
not overwrite possible changes made by users.

This fixes the following lintian warnings:
W: lcs-security: file-in-unusual-dir home/rssh/.ssh/id_rsa
W: lcs-security: file-in-unusual-dir home/rssh/.ssh/id_rsa.pub
W: lcs-security: file-in-unusual-dir root/.bashrc

The long description of the warning is:
N:
N:   This file or symbolic link is in a directory where files are not
N:   normally installed by Debian packages.
N:
N:   Severity: normal, Certainty: certain
N:
N:   Check: files, Type: binary, udeb
N: