Cleanup maintainer scripts (postinst, postrm)
Detailed changes:
* drop dead code (unused variables)
* fixup postrm

The postrm script restarts fail2ban and calls lcs-make-sudoers to
regenerate the sudoers file. Those actions were done in all cases
but it only makes sense to do them when the package is purged. It's
only in this case that the /etc/fail2ban/filter.d/* and /etc/lcs/sudo.d/*
files are actually removed. So I restrict the actions to this specific
case.

Also the actions were not protected against failures with "|| true"
but the "postrm purge" step has no guaranty that the dependencies are
available. When the dependencies are missing the postrm script would
fail and the package cannot be purged. This is now fixed too.