Ignore binary-without-manpage lintian warning.
The Debian policy requires that each program available on the PATH
has a corresponding manual page. The LCS adds scripts that do not have
manual pages. Since none of those programs are for general use but are
only used by other LCS components, it's not a big deal.

Thus we add a lintian override to ignore this warning.