Fix shebang of lcs-remove-sudoers
The shebang contained an unwanted trailing slash ("#!/bin/bash/" instead
of "#!/bin/bash"). At the same time the script does not need any
Bash-specific feature and can thus be executed by /bin/sh. So switch
the shebang to that shell instead.

This fixes the following lintian warning:
W: lcs-preinstall: unusual-interpreter usr/sbin/lcs-remove-sudoers #!/bin/bash/
N:
N:   This package contains a script for an interpreter that is not shipped
N:   in the package and is not known to Lintian. It is possible that there
N:   is a typo or the interpreter is not executable. If not, please file a
N:   wishlist bug against lintian, so it can be added to the list of known
N:   interpreters.
N:
N:   Severity: normal, Certainty: possible
N:
N:   Check: scripts, Type: binary
N: