Keep a single init script and install it with dh_installinit
The package contained two init scripts: etc/init.d/lcs-sftp and
debian/init.d. Only the former was used and the init script was
not properly registered in the maintainer scripts.

Move everything related to the init script under debian and rely
on dh_installinit to add the appropriate update-rc.d calls.