lcs-preinstall: provide /etc/ssh/sshd_config_sftp as a configuration file
This file is specific to lcs-preinstall, there's no reason to create it
dynamically instead of shipping it as a proper configuration file. Also
get rid of the manual update-rc.d call in lcs-sshd-config.

This ensures that local modifications are preserved and delegates
more work to dpkg/dh_installinit instead of a relying on a LCS script.