lcs-preinstall: cleanup prerm script with respect to possible parameters
The prerm script was written as if it accepted the same parameters than
postrm but it's not the case. Fix the comments with a real list of
parameters (copied from /var/lib/dpkg/info/dpkg.prerm).

The exact ways that maintainer scripts can be called are also documented
at the following places:
https://www.debian.org/doc/debian-policy/ch-maintainerscripts.html#s-mscriptsinstact
https://wiki.debian.org/MaintainerScripts
https://people.debian.org/~srivasta/MaintainerScripts.html

We want to disable the commands in /etc/sudoers only on package removal so
call lcs-remove-sudoers only when "prerm remove" is called.