Bump Standards-Version to 3.9.4.
The requirement of the Debian policy evolve over time. The current version
of the Debian policy that the package is known to follow is documented
in the Standards-Version field. From time to time, one is supposed
to check /usr/share/doc/debian-policy/upgrading-checklist.txt.gz to see
the changes made since the last time the package was checked against
policy. And then update the package accordingly. And then change the
version in Standards-Version.

With the cleanups done we are reasonably clean at the lintian level
and we can pretend that we are in line with the requirements of Debian
policy as they were set for Debian 7 Wheezy. Thus we bump the
Standards-Version field to 3.9.4 for all the packages we cleaned up.