lcs-preinstall: fix lcs-sftp init script to create /var/run/sshd, the PrivSep empty dir
This directory must be shared between ssh and lcs-sftp as the directory
is hardcoded in the sshd executable: it has been set at compilation
time with the ./configure option “--with-privsep-path=/var/run/sshd”.

This is not a problem as the role of this directory is to be empty
so that a chroot into it doesn't give access to anything on the system.
sshd uses this in some parts of its work to reduce the privileges
and limit the damages in case of a security bug.

Hopefully this change will fix the issue with the lcs-sftp service
not starting at boot.